TabbedPane: use float arc for tab area button background

+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatTabbedPaneUI.java
@@ -1559,7 +1559,7 @@ public class FlatTabbedPaneUI
 			FlatUIUtils.paintComponentBackground( g, left, top,
 				getWidth() - left - right,
 				getHeight() - top - bottom,
-				0, scale( buttonArc ) );
+				0, scale( (float) buttonArc ) );
 		}
 	}
